Changed: Room and Floor
As of right now, the bug for the speed is fixed. Rooms are recreated
each time without changing the rooms layout.

Room has two constructors at the moment. One is for a brand new room,
the other is for when the player is traversing to a room.

Will work on the new level designs soon.